Fix : 검색 > sort 변경시 url도 변경되도록 수정
